+++ docs/release-notes/change-log.md
@@ -9,9 +9,9 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## Unreleased
+## v1.10.5
 
-* Fixed an(other) issue where multiple stores running on the same stage with different initialBlocks will fail to proress
+* Fixed an(other) issue where multiple stores running on the same stage with different initialBlocks will fail to proress (and hang)
 
 ## v1.10.4